Address comments from review.
Mainly refactor MockResponseStream to bigtable/client/testing.
