+++ bigtable/CMakeLists.txt
@@ -132,6 +132,7 @@ target_link_libraries(bigtable_client
 add_library(bigtable_client_testing
     client/testing/chrono_literals.h
     client/testing/mock_data_client.h
+    client/testing/mock_response_stream.h
     client/testing/table_test_fixture.h
     client/testing/table_test_fixture.cc)
 target_link_libraries(bigtable_client_testing
