+++ bigtable/client/table_readrow_test.cc
@@ -15,19 +15,9 @@
 #include "bigtable/client/table.h"
 #include "bigtable/client/testing/table_test_fixture.h"
 
-namespace {
-class MockResponseStream : public grpc::ClientReaderInterface<
-                               ::google::bigtable::v2::ReadRowsResponse> {
- public:
-  MOCK_METHOD0(WaitForInitialMetadata, void());
-  MOCK_METHOD0(Finish, grpc::Status());
-  MOCK_METHOD1(NextMessageSize, bool(std::uint32_t *));
-  MOCK_METHOD1(Read, bool(::google::bigtable::v2::ReadRowsResponse *));
-};
-
 /// Define helper types and functions for this test.
+namespace {
 class TableReadRowTest : public bigtable::testing::TableTestFixture {};
-
 }  // anonymous namespace
 
 TEST_F(TableReadRowTest, ReadRowSimple) {
@@ -45,8 +35,7 @@ TEST_F(TableReadRowTest, ReadRowSimple) {
       }
 )");
 
-  // must be a new pointer, it is wrapped in unique_ptr by ReadRows
-  auto stream = absl::make_unique<MockResponseStream>();
+  auto stream = absl::make_unique<bigtable::testing::MockResponseStream>();
   EXPECT_CALL(*stream, Read(_))
       .WillOnce(Invoke([&response](btproto::ReadRowsResponse *r) {
         *r = response;
@@ -75,8 +64,7 @@ TEST_F(TableReadRowTest, ReadRowMissing) {
   using namespace ::testing;
   namespace btproto = ::google::bigtable::v2;
 
-  // must be a new pointer, it is wrapped in unique_ptr by ReadRows
-  auto stream = absl::make_unique<MockResponseStream>();
+  auto stream = absl::make_unique<bigtable::testing::MockResponseStream>();
   EXPECT_CALL(*stream, Read(_)).WillOnce(Return(false));
   EXPECT_CALL(*stream, Finish()).WillOnce(Return(grpc::Status::OK));
